Add Kokkos support for accelerators (#327)
This PR adds Kokkos accelerator backend as an alternative to the native ParFlow CUDA backend to support more accelerator devices. The implementation does not rely on any CUDA-specific arguments but still requires Unified Memory support from the accelerator devices. It should be compatible with AMD GPUs when sufficient Unified Memory support is available from this vendor.

The performance of using CUDA through the Kokkos library is slightly worse in comparison to the ParFlow native CUDA implementation. This is because a general Kokkos implementation cannot leverage certain CUDA features such as cudaMemset() for initializations or CUDA pinned host/device memory for MPI buffers. Also, Kokkos determines grid and block sizes for compute kernels differently.

The RMM pool allocator for Unified Memory can be used with Kokkos (when using Kokkos CUDA backend) and improves the performance very significantly. In the future, a Unified Memory pool allocator that supports AMD cards is likely needed to achieve good performance.

User instructions on how to use Kokkos backend can be found from README-GPU.md. Basically, the following must be specified for CMake when compiling ParFlow:

-DPARFLOW_ACCELERATOR_BACKEND=kokkos -DKokkos_ROOT=/path/to/Kokkos

Furthermore, the performance of the simulation initialization phase has been improved significantly when using GPUs (with CUDA and Kokkos).

Co-authored-by: Steven Smith <smith84@llnl.gov>