Fix bug in run.dist() (#536)
run.dist() does not re-distribute well files that have been already been
distributed in a way that p, q, r do not exactly divide NX, NY, NZ. The
problem lies in trying to "estimate" p, q, r when reading those
distributed files in with the ParflowBinaryReader.

This pull request removes the precomputation option (which cannot work
if we allow distributing grids unevenly), and adds explicit calculation
for p, q, r in the read_subgrid_info. Also, a test for
ParflowBinaryReader is added, which distributes and redistributes a pfb
file. That test fails with the previous implementation.

---------

Co-authored-by: gartavanis <gartavanis@users.noreply.github.com>
Co-authored-by: Steven Smith <smith84@llnl.gov>