Check parflow inputs for NaNs (#557)
This PR adds a new method check_nans() to the Run class of pftools. This
checks the static and forcing files for NaNs. The aim is to expand the
suite with other input checks, e.g. mask checks, values that are
out-of-range etc, depending on the type of simulation (box or solid
domain). This can eventually be integrated into Run.run(), so that the
checks are executed automatically before the simulation starts.

Also, the LW test.py test is modified to check its inputs.

---------

Co-authored-by: gartavanis <gartavanis@users.noreply.github.com>
Co-authored-by: George Artavanis <ga6@verde.princeton.edu>