Implement a new overland flow formulation with channel width (#535)
Added reading channel width X and Y files. The files are not yet used,
this change is in preparation for using in a pending overland flow
formulation change that includes channel width as a parameter. Test
added to ensure that we could read both constant channel width data and
channel width data from PFBs.

---------

Co-authored-by: gartavanis <32808713+gartavanis@users.noreply.github.com>
Co-authored-by: Steven Smith <smith84@llnl.gov>
Co-authored-by: George Artavanis <ga6@verde.princeton.edu>