Update overlandflow_eval_Kin.c (#603)
Fix to overlandflow_eval_Kin to correct issue with overland surface boundary and Dirichlet side boundary.

The issue is with the “top” pressure value. The top function is used here
[k1y = pfmax((int)top_dat[itop + sy_v],0);](https://github.com/parflow/parflow/blob/483aac5603f6ad69f04e1a12faf5f84859ff5ea6/pfsimulator/parflow_lib/overlandflow_eval_Kin.c#L139)
[to get the upwinded pressure](https://github.com/parflow/parflow/blob/483aac5603f6ad69f04e1a12faf5f84859ff5ea6/pfsimulator/parflow_lib/overlandflow_eval_Kin.c#L156).

The top value is zero and we are counting on the pressures outside the active domain being zero. I actually don’t think they always are, I think the BC information gets inserted into the inactive cells adjacent to the boundary and this value at k=0 is what’s being used any time the slopes point into the domain. This fix checks for the upper right and
upper upper top values, if they are <0 (meaning outside the active domain) it assigns an upper pressure of zero, otherwise it assigns the pressure from the upper or right cells (north and east cells) as normal.

---------

Co-authored-by: Steven Smith <smith84@llnl.gov>